Fixed alpha-beta.

+++ src/main/java/com/igorternyuk/engine/player/ai/BoardEvaluator.java
@@ -6,5 +6,5 @@ import com.igorternyuk.engine.board.Board;
  * Created by igor on 06.12.18.
  */
 public interface BoardEvaluator {
-    int evaluate(Board board, int depth);
+    double evaluate(Board board, int depth);
 }
